+++ docs/sphinx/source/conf.py
@@ -79,7 +79,6 @@ typehints_document_rtype = True
 intersphinx_mapping = {
     'python': ('https://docs.python.org/3/', None),
     'numpy': ('https://numpy.org/doc/stable/', None),
-    'matplotlib': ('https://matplotlib.org/stable/', None),
     'torch': ('https://pytorch.org/docs/stable/', None),
 }