+++ src/vcs/_visualize_vcs/_global_nas/__init__.py
@@ -1,3 +1,0 @@
-from .visualizer import visualize_global_nas
-
-__all__ = ["visualize_global_nas"]