+++ src/vcs/_visualize_vcs/_local_sas/__init__.py
@@ -1,3 +1,0 @@
-from .visualizer import visualize_las, visualize_las_load_sharing
-
-__all__ = ["visualize_las", "visualize_las_load_sharing"]